+++ HttpAppRequest.js
@@ -10,8 +10,11 @@ function HttpAppRequest ( app, req, res ) {
 	this.Domain.add( req );
 	this.Domain.add( res );
 
-	this.Response.once( 'close', this.dispose.bind( this ) );
-	this.Domain.on( 'error', this.onError.bind( this ) );
+	this._onResponseClose = this.dispose.bind( this );
+	this._onDomainError = this.onError.bind( this );
+
+	this.Response.once( 'close', this._onResponseClose );
+	this.Domain.on( 'error', this._onDomainError );
 	
 	this.Domain.run( this.onHttpHeaders.bind( this ) );
 }
@@ -21,6 +24,7 @@ HttpAppRequest.define( {
 	dispose: function () {
 		if ( this.Domain ) {
 			this.App.unregisterRequest( this );
+			this.Domain.removeListener( 'error', this._onDomainError );
 			this.Domain.remove( this.Request );
 			this.Domain.remove( this.Response );
 			this.Domain = null;
@@ -47,7 +51,7 @@ HttpAppRequest.define( {
 
 	onError: function ( err ) {
 		console.error( err.message, err.stack );
-		this.App.close( 1 );
+		this.App.close();
 	},
 
 } );